+++ src/app/services/control-api/control-api.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ControlApiService {
+
+  constructor() { }
+}
